Incorrect word usage
Line  186- "The calls to act() will also flush the effects inside of them." uses 'inside of'.
Here, "inside" as a preposition meaning "within the interior of something" would be correct. The phrase the 'inside of X' uses inside as a noun, meaning "interior or inner part". Eg.: "The inside of each call...", but just saying "The calls to act() will also flush the effects inside them" is both correct and direct.